requirements.txt
flask
gunicorn
pokemontcgsdk
python-dotenv
requests
